+++ scripts/datasets/ilsvrc_vid.py
@@ -3,8 +3,7 @@ which is Object detection from video in Large Scale Visual
 Recognition Challenge 2015 (ILSVRC2015)
 Code adapted from https://github.com/STVIR/pysot"""
 import json
-from os.path import join, isdir, expanduser
-from os import listdir, chdir, system
+import os
 import glob
 from concurrent import futures
 import time
@@ -14,10 +13,10 @@ try:
 except ImportError:
     import xml.etree.ElementTree as ET
 import tarfile
-import sys
 import numpy as np
 from gluoncv.utils import download, makedirs
 from gluoncv.utils.filesystem import try_import_cv2
+from gluoncv.utils.data.tracking import crop_like_SiamFC, printProgress
 
 def parse_args():
     """VID dataset parameter."""
@@ -30,16 +29,16 @@ def parse_args():
     parser.add_argument('--num-threads', type=int, default=12,
                         help='threads number')
     args = parser.parse_args()
-    args.download_dir = expanduser(args.download_dir)
+    args.download_dir = os.path.expanduser(args.download_dir)
     return args
 
-def download_VID(overwrite=False):
+def download_VID(args, overwrite=False):
     """download VID dataset and Unzip to download_dir"""
     _DOWNLOAD_URLS = [ 
     ('http://bvisionweb1.cs.unc.edu/ilsvrc2015/ILSVRC2015_VID.tar.gz',
     '077dbdea4dff1853edd81b04fa98e19392287ca3'),
     ]
-    if not isdir(args.download_dir):
+    if not os.path.isdir(args.download_dir):
         makedirs(args.download_dir)
     for url, checksum in _DOWNLOAD_URLS:
         filename = download(url, path=args.download_dir, overwrite=overwrite, sha1_hash=checksum)
@@ -58,16 +57,16 @@ def parse_vid(ann_base_path, args):
     sub_sets = sorted({'a', 'b', 'c', 'd', 'e'})
     vid = []
     for sub_set in sub_sets:
-        sub_set_base_path = join(ann_base_path, sub_set)
-        videos = sorted(listdir(sub_set_base_path))
+        sub_set_base_path = os.path.join(ann_base_path, sub_set)
+        videos = sorted(os.listdir(sub_set_base_path))
         s = []
         for vi, video in enumerate(videos):
             print('subset: {} video id: {:04d} / {:04d}'.format(sub_set, vi, len(videos)))
             v = dict()
-            v['base_path'] = join(sub_set, video)
+            v['base_path'] = os.path.join(sub_set, video)
             v['frame'] = []
-            video_base_path = join(sub_set_base_path, video)
-            xmls = sorted(glob.glob(join(video_base_path, '*.xml')))
+            video_base_path = os.path.join(sub_set_base_path, video)
+            xmls = sorted(glob.glob(os.path.join(video_base_path, '*.xml')))
             for xml in xmls:
                 f = dict()
                 xmltree = ET.parse(xml)
@@ -94,102 +93,8 @@ def parse_vid(ann_base_path, args):
             s.append(v)
         vid.append(s)
     print('save json (raw vid info), please wait 1 min~')
-    json.dump(vid, open(join(args.download_dir, 'vid.json'), 'w'), indent=4, sort_keys=True)
+    json.dump(vid, open(os.path.join(args.download_dir, 'vid.json'), 'w'), indent=4, sort_keys=True)
     print('done!')
-
-def printProgress(iteration, total, prefix='', suffix='', decimals=1, barLength=100):
-    """
-    Call in a loop to create terminal progress bar
-
-    Parameters
-    ----------
-        iteration: int : current iteration.
-        total: int, total iterations.
-        prefix: str, prefix string.
-        suffix: str, suffix string.
-        decimals: int, positive number of decimals in percent complete.
-        barLength: int, character length of bar.
-    """
-    formatStr = "{0:." + str(decimals) + "f}"
-    percents = formatStr.format(100 * (iteration / float(total)))
-    filledLength = int(round(barLength * iteration / float(total)))
-    bar = '' * filledLength + '-' * (barLength - filledLength)
-    sys.stdout.write('\r%s |%s| %s%s %s' % (prefix, bar, percents, '%', suffix)),
-    if iteration == total:
-        sys.stdout.write('\x1b[2K\r')
-    sys.stdout.flush()
-
-def crop_hwc(image, bbox, out_sz, padding=(0, 0, 0)):
-    """
-    crop image
-
-    Parameters
-    ----------
-    image: np.array, image
-    bbox: np or list, bbox coordinate [xmin,ymin,xmax,ymax]
-    out_sz: int , crop image size
-
-    Return:
-        crop result
-    """
-    a = (out_sz-1) / (bbox[2]-bbox[0])
-    b = (out_sz-1) / (bbox[3]-bbox[1])
-    c = -a * bbox[0]
-    d = -b * bbox[1]
-    mapping = np.array([[a, 0, c],
-                        [0, b, d]]).astype(np.float)
-    cv2 = try_import_cv2()
-    crop = cv2.warpAffine(image, mapping, (out_sz, out_sz), borderMode=cv2.BORDER_CONSTANT, borderValue=padding)
-    return crop
-
-
-def pos_s_2_bbox(pos, s):
-    """
-    from center_x,center_y,s to get bbox
-
-    Parameters
-    ----------
-    pos , x, bbox
-    s , int, bbox size
-
-    Return:
-        [x_min,y_min,x_max,y_max]
-    """
-    return [pos[0]-s/2, pos[1]-s/2, pos[0]+s/2, pos[1]+s/2]
-
-
-def crop_like_SiamFC(image, bbox, context_amount=0.5, exemplar_size=127, instance_size=255, padding=(0, 0, 0)):
-    """
-    Dataset curation and avoid image resizing during training
-    if the tight bounding box has size (w, h) and the context margin is p,
-    then the scale factor s is chosen such that the area of the scaled rectangle is equal to a constant
-    s(w+2p)×s(h+2p)=A.
-
-    Parameters
-    ----------
-    image: np.array, image
-    bbox: list or np.array, bbox
-    context_amount: float, the amount of context to be half of the mean dimension
-    exemplar_size: int, exemplar_size
-    instance_size: int, instance_size
-
-    Return:
-        crop result exemplar z and instance x
-    """
-    target_pos = [(bbox[2]+bbox[0])/2., (bbox[3]+bbox[1])/2.]
-    target_size = [bbox[2]-bbox[0], bbox[3]-bbox[1]]
-    wc_z = target_size[1] + context_amount * sum(target_size)
-    hc_z = target_size[0] + context_amount * sum(target_size)
-    s_z = np.sqrt(wc_z * hc_z)
-    scale_z = exemplar_size / s_z
-    d_search = (instance_size - exemplar_size) / 2
-    pad = d_search / scale_z
-    s_x = s_z + 2 * pad
-
-    z = crop_hwc(image, pos_s_2_bbox(target_pos, s_z), exemplar_size, padding)
-    x = crop_hwc(image, pos_s_2_bbox(target_pos, s_x), instance_size, padding)
-    return z, x
-
 
 def crop_video(args, sub_set, video, crop_path, ann_base_path):
     """
@@ -203,17 +108,16 @@ def crop_video(args, sub_set, video, crop_path, ann_base_path):
     ann_base_path: str, Annotations base path
     """
     cv2 = try_import_cv2()
-    video_crop_base_path = join(crop_path, sub_set, video)
-    if not isdir(video_crop_base_path):
+    video_crop_base_path = os.path.join(crop_path, sub_set, video)
+    if not os.path.isdir(video_crop_base_path):
         makedirs(video_crop_base_path)
-    sub_set_base_path = join(ann_base_path, sub_set)
-    xmls = sorted(glob.glob(join(sub_set_base_path, video, '*.xml')))
+    sub_set_base_path = os.path.join(ann_base_path, sub_set)
+    xmls = sorted(glob.glob(os.path.join(sub_set_base_path, video, '*.xml')))
     for xml in xmls:
         xmltree = ET.parse(xml)
         objects = xmltree.findall('object')
         objs = []
         filename = xmltree.findall('filename')[0].text
-        # print(xml.replace('xml', 'JPEG').replace('Annotations', 'Data'))
         im = cv2.imread(xml.replace('xml', 'JPEG').replace('Annotations', 'Data'))
         avg_chans = np.mean(im, axis=(0, 1))
         for object_iter in objects:
@@ -222,24 +126,23 @@ def crop_video(args, sub_set, video, crop_path, ann_base_path):
             bbox = [int(bndbox.find('xmin').text), int(bndbox.find('ymin').text),
                     int(bndbox.find('xmax').text), int(bndbox.find('ymax').text)]
             z, x = crop_like_SiamFC(im, bbox, instance_size=args.instance_size, padding=avg_chans)
-            cv2.imwrite(join(args.download_dir, video_crop_base_path, '{:06d}.{:02d}.z.jpg'.format(int(filename), trackid)), z)
-            cv2.imwrite(join(args.download_dir, video_crop_base_path, '{:06d}.{:02d}.x.jpg'.format(int(filename), trackid)), x)
+            cv2.imwrite(os.path.join(args.download_dir, video_crop_base_path, '{:06d}.{:02d}.z.jpg'.format(int(filename), trackid)), z)
+            cv2.imwrite(os.path.join(args.download_dir, video_crop_base_path, '{:06d}.{:02d}.x.jpg'.format(int(filename), trackid)), x)
 
 def par_crop(args, ann_base_path):
     """
     Dataset curation, crop data and transform the format of label
-
     Parameters
     ----------
     ann_base_path: str, Annotations base path
     """
-    crop_path = join(args.download_dir, './crop{:d}'.format(int(args.instance_size)))
-    if not isdir(crop_path):
+    crop_path = os.path.join(args.download_dir, './crop{:d}'.format(int(args.instance_size)))
+    if not os.path.isdir(crop_path):
         makedirs(crop_path)
     sub_sets = sorted({'a', 'b', 'c', 'd', 'e'})
     for sub_set in sub_sets:
-        sub_set_base_path = join(ann_base_path, sub_set)
-        videos = sorted(listdir(sub_set_base_path))
+        sub_set_base_path = os.path.join(ann_base_path, sub_set)
+        videos = sorted(os.listdir(sub_set_base_path))
         n_videos = len(videos)
         with futures.ProcessPoolExecutor(max_workers=args.num_threads) as executor:
             fs = [executor.submit(crop_video, args, sub_set, video, crop_path, ann_base_path) for video in videos]
@@ -247,10 +150,11 @@ def par_crop(args, ann_base_path):
                 # Write progress to error so that it can be seen
                 printProgress(i, n_videos, prefix=sub_set, suffix='Done ', barLength=40)
 
+
 def gen_json(args):
     """transform json and generate train and val json, prepare for tracking dataloader"""
     print('load json (raw vid info), please wait 20 seconds~')
-    vid = json.load(open(join(args.download_dir, 'vid.json'), 'r'))
+    vid = json.load(open(os.path.join(args.download_dir, 'vid.json'), 'r'))
     snippets = dict()
     n_snippets = 0
     n_videos = 0
@@ -300,26 +204,31 @@ def gen_json(args):
     print('done!')
 
 def symlink(args):
-    """Soft connection"""
-    chdir(args.download_dir)
-    system('ln -sfb $PWD/ILSVRC2015/Annotations/VID/train/ILSVRC2015_VID_train_0000 ILSVRC2015/Annotations/VID/train/a')
-    system('ln -sfb $PWD/ILSVRC2015/Annotations/VID/train/ILSVRC2015_VID_train_0001 ILSVRC2015/Annotations/VID/train/b')
-    system('ln -sfb $PWD/ILSVRC2015/Annotations/VID/train/ILSVRC2015_VID_train_0002 ILSVRC2015/Annotations/VID/train/c')
-    system('ln -sfb $PWD/ILSVRC2015/Annotations/VID/train/ILSVRC2015_VID_train_0003 ILSVRC2015/Annotations/VID/train/d')
-    system('ln -sfb $PWD/ILSVRC2015/Annotations/VID/val ILSVRC2015/Annotations/VID/train/e')
+    """Soft connection in VID dataset """
+    def per_symlink(src, dst):
+        """Soft connection"""
+        src = os.path.join(args.download_dir, src)
+        dst = os.path.join(args.download_dir, dst)
+        if not os.path.isdir(dst):
+            os.symlink(src, dst)
+    per_symlink('ILSVRC2015/Annotations/VID/train/ILSVRC2015_VID_train_0000', 'ILSVRC2015/Annotations/VID/train/a')
+    per_symlink('ILSVRC2015/Annotations/VID/train/ILSVRC2015_VID_train_0001', 'ILSVRC2015/Annotations/VID/train/b')
+    per_symlink('ILSVRC2015/Annotations/VID/train/ILSVRC2015_VID_train_0002', 'ILSVRC2015/Annotations/VID/train/c')
+    per_symlink('ILSVRC2015/Annotations/VID/train/ILSVRC2015_VID_train_0003', 'ILSVRC2015/Annotations/VID/train/d')
+    per_symlink('ILSVRC2015/Annotations/VID/val', 'ILSVRC2015/Annotations/VID/train/e')
 
-    system('ln -sfb $PWD/ILSVRC2015/Data/VID/train/ILSVRC2015_VID_train_0000 ILSVRC2015/Data/VID/train/a')
-    system('ln -sfb $PWD/ILSVRC2015/Data/VID/train/ILSVRC2015_VID_train_0001 ILSVRC2015/Data/VID/train/b')
-    system('ln -sfb $PWD/ILSVRC2015/Data/VID/train/ILSVRC2015_VID_train_0002 ILSVRC2015/Data/VID/train/c')
-    system('ln -sfb $PWD/ILSVRC2015/Data/VID/train/ILSVRC2015_VID_train_0003 ILSVRC2015/Data/VID/train/d')
-    system('ln -sfb $PWD/ILSVRC2015/Data/VID/val ILSVRC2015/Data/VID/train/e')
+    per_symlink('ILSVRC2015/Data/VID/train/ILSVRC2015_VID_train_0000', 'ILSVRC2015/Data/VID/train/a')
+    per_symlink('ILSVRC2015/Data/VID/train/ILSVRC2015_VID_train_0001', 'ILSVRC2015/Data/VID/train/b')
+    per_symlink('ILSVRC2015/Data/VID/train/ILSVRC2015_VID_train_0002', 'ILSVRC2015/Data/VID/train/c')
+    per_symlink('ILSVRC2015/Data/VID/train/ILSVRC2015_VID_train_0003', 'ILSVRC2015/Data/VID/train/d')
+    per_symlink('ILSVRC2015/Data/VID/val', 'ILSVRC2015/Data/VID/train/e')
 
 def main(args):
     # download VID dataset
     download_VID(args)
     print('VID dataset has already download completed')
-    VID_base_path = join(args.download_dir, 'ILSVRC2015')
-    ann_base_path = join(VID_base_path, 'Annotations/VID/train/')
+    VID_base_path = os.path.join(args.download_dir, 'ILSVRC2015')
+    ann_base_path = os.path.join(VID_base_path, 'Annotations/VID/train/')
     symlink(args)
     # Format XML and save it in JSON
     parse_vid(ann_base_path, args)
